refactor(rust): move pipe tests into a separate module
